action/pushing/youtube_action_module/requirements.txt
# Web framework
quart==0.19.4
hypercorn==0.16.0

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# Database
pydal==20231105.1

# Google APIs
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# JWT Authentication
pyjwt==2.8.0

# PostgreSQL support
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
